cb_get_articles_id/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
google-cloud-storage
watchdog==4.*
db-dtypes
scipy
